Refactor "stockquotes" to "yahoostockdata"
Refactor "stockquotes" to "yahoostockdata". This is because YQL no
longer does keystats data (its broken) and I've been come distrustful of
Yahoo as a sole data source. So I'm preparing to move to allowing
multiple data sources. So I'm going to reduce (in a future commit) how
many functions there are and just get basic quote and dividend info
(which both still work).